leaderBoards/pyproject.toml
[project]
name = "leaderboards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "aiosignal>=1.3.2",
    "annotated-types>=0.7.0",
    "anyio>=4.8.0",
    "asttokens>=3.0.0",
    "attrs>=24.3.0",
    "bcrypt>=4.2.1",
    "certifi>=2024.12.14",
    "cffi>=1.17.1",
    "click>=8.1.8",
    "colorama>=0.4.6",
    "comm>=0.2.2",
    "cryptography>=44.0.0",
    "debugpy>=1.8.11",
    "decorator>=5.1.1",
    "deprecation>=2.1.0",
    "ecdsa>=0.19.0",
    "et-xmlfile>=2.0.0",
    "executing>=2.1.0",
    "fastapi>=0.115.6",
    "frozenlist>=1.5.0",
    "gotrue>=2.11.1",
    "h11>=0.14.0",
    "h2>=4.1.0",
    "hpack>=4.0.0",
    "httpcore>=1.0.7",
    "httpx>=0.27.2",
    "hyperframe>=6.0.1",
    "idna>=3.10",
    "ipykernel>=6.29.5",
    "ipython>=8.31.0",
    "jedi>=0.19.2",
    "jinja2>=3.1.5",
    "jupyter-client>=8.6.3",
    "jupyter-core>=5.7.2",
    "jwcrypto>=1.5.6",
    "jwt>=1.3.1",
    "markupsafe>=3.0.2",
    "matplotlib-inline>=0.1.7",
    "multidict>=6.1.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.1",
    "openpyxl>=3.1.5",
    "packaging>=24.2",
    "pandas>=2.2.3",
    "parso>=0.8.4",
    "passlib>=1.7.4",
    "platformdirs>=4.3.6",
    "postgrest>=0.19.1",
    "prompt-toolkit>=3.0.48",
    "propcache>=0.2.1",
    "psutil>=6.1.1",
    "pure-eval>=0.2.3",
    "pyasn1>=0.6.1",
    "pycparser>=2.22",
    "pydantic>=2.10.5",
    "pydantic-core>=2.27.2",
    "pygments>=2.19.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-jwt>=4.1.0",
    "python-multipart>=0.0.20",
    "pytz>=2024.2",
    "pyzmq>=26.2.0",
    "realtime>=2.1.0",
    "rsa>=4.9",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "stack-data>=0.6.3",
    "starlette>=0.41.3",
    "storage3>=0.11.0",
    "supabase>=2.11.0",
    "supafunc>=0.9.0",
    "tornado>=6.4.2",
    "traitlets>=5.14.3",
    "typing-extensions>=4.12.2",
    "tzdata>=2024.2",
    "uvicorn>=0.34.0",
    "wcwidth>=0.2.13",
    "websockets>=13.1",
    "yarl>=1.18.3",
]
